+++ js/mt_Device.js
@@ -10,9 +10,10 @@ DO NOT REMOVE THIS COPYRIGHT
  WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 */
 
-import * as mt_MMS from "./mt_mms.js";
+import * as mt_MMS from "./API_mmsHID.js";
 import * as mt_HID from "./mt_hid.js";
-import * as mt_MQTT from "./mmsMQTT_API.js";
+import * as mt_MQTT from "./API_mmsMQTT.js";
+import * as mt_WSS from "./API_mmsWebSocket.js";
 
 var _Transport = "HID";
 var _Type = "MMS";
@@ -32,20 +33,19 @@ export function setType(name) {
     _Type = name;
 };
 
-
 export async function OpenDevice() 
 {
     switch (`${_Type.toLowerCase()}_${_Transport.toLowerCase()}`) {
         case "mms_wss":
-            
+            return await mt_WSS.OpenWS();
             break;
         case "mms_mqtt":
-            await mt_MQTT.OpenMQTT();
+            return await mt_MQTT.OpenMQTT();
             break;
         case "mms_hid":
-            await mt_HID.openDevice();;
+            return await mt_HID.openDevice();;
             break;
-        default: 
+        default:             
             break;
     }
 };
@@ -54,12 +54,13 @@ export async function CloseDevice()
 {
     switch (`${_Type.toLowerCase()}_${_Transport.toLowerCase()}`) {
         case "mms_wss":
+            return await mt_WSS.CloseWS();
             break;
         case "mms_mqtt":
-            await mt_MQTT.CloseMQTT();
+            return await mt_MQTT.CloseMQTT();
             break;
         case "mms_hid":
-            await mt_HID.closeDevice();
+            return await mt_HID.closeDevice();
             break;
         default: 
             break;
@@ -70,15 +71,17 @@ export async function SendCommand(command)
 {
 switch (_Transport.toLowerCase()) {
     case "mms_wss":
+        return await mt_WSS.SendCommand(command);        
         break;
     case "mms_mqtt":
+        return await mt_MQTT.SendCommand(command);
         break;
     case "mms_hid":
+        return await mt_HID.SendCommand(command);
         break;
-    case "mms_ble":
         break;
     default: 
-        
+
         break;
 }
 };
