+++ js/mt_rms_v5.js
@@ -13,7 +13,7 @@ DO NOT REMOVE THIS COPYRIGHT
 "use strict";
 import * as mt_V5 from "./mt_v5.js";
 import * as mt_Utils from "./mt_utils.js";
-import * as mt_RMS_API from "./mt_rms_api.js";
+import * as mt_RMS_API from "./API_rms.js";
 
 let _KSN = "";
 let _UIK = "";
